arm: Move ArmMemoryInterface up one module level
