+++ probe-rs/src/architecture/arm/core/armv8a.rs
@@ -10,8 +10,8 @@ use super::{
 };
 use crate::{
     architecture::arm::{
-        core::armv8a_debug_regs::*, memory::adi_v5_memory_interface::ArmMemoryInterface,
-        sequences::ArmDebugSequence, ArmError,
+        core::armv8a_debug_regs::*, memory::ArmMemoryInterface, sequences::ArmDebugSequence,
+        ArmError,
     },
     core::{
         memory_mapped_registers::MemoryMappedRegister, CoreRegisters, RegisterId, RegisterValue,
